box: consolidate all read view code in one module
Historically, the read view implementation is partially located in
the schema and misc box modules. Now, after we imported the main read
view module from the Tarantool EE repository, we can consolidate all
the bits in one place.

Follow-up #12912

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring
