+++ src/box/lua/misc.cc
@@ -42,7 +42,6 @@
 #include "box/lua/tuple.h"
 #include "box/memtx_tx.h"
 #include "box/port.h"
-#include "box/read_view.h"
 #include "box/space_cache.h"
 #include "box/tuple.h"
 #include "box/txn.h"
@@ -521,68 +520,6 @@ lbox_memtx_tx_gc(struct lua_State *L)
 	return 0;
 }
 
-/** {{{ Read view utils. **/
-
-void
-lbox_push_read_view(struct lua_State *L, const struct read_view *rv)
-{
-	lua_newtable(L);
-	luaL_pushuint64(L, rv->id);
-	lua_setfield(L, -2, "id");
-	lua_pushstring(L, rv->name);
-	lua_setfield(L, -2, "name");
-	lua_pushboolean(L, rv->is_system);
-	lua_setfield(L, -2, "is_system");
-	lua_pushnumber(L, rv->timestamp);
-	lua_setfield(L, -2, "timestamp");
-	luaT_pushvclock(L, &rv->vclock);
-	lua_setfield(L, -2, "vclock");
-	luaL_pushint64(L, vclock_sum(&rv->vclock));
-	lua_setfield(L, -2, "signature");
-}
-
-static bool
-lbox_read_view_list_cb(struct read_view *rv, void *arg)
-{
-	struct lua_State *L = (struct lua_State *)arg;
-	assert(lua_gettop(L) >= 1);
-	assert(lua_type(L, -1) == LUA_TTABLE);
-	lbox_push_read_view(L, rv);
-	lua_rawseti(L, -2, lua_objlen(L, -2) + 1);
-	return true;
-}
-
-/**
- * Pushes an unsored array of all open read views to the Lua stack.
- * Each read view is represented by a plain Lua table.
- */
-static int
-lbox_read_view_list(struct lua_State *L)
-{
-	lua_newtable(L);
-	read_view_foreach(lbox_read_view_list_cb, L);
-	return 1;
-}
-
-/**
- * Given a read view object (a table that has the 'id' field), pushes
- * the read view status string ('open' or 'closed') to the Lua stack.
- */
-static int
-lbox_read_view_status(struct lua_State *L)
-{
-	lua_getfield(L, 1, "id");
-	uint64_t id = luaL_checkuint64(L, -1);
-	struct read_view *rv = read_view_by_id(id);
-	if (rv == NULL)
-		lua_pushliteral(L, "closed");
-	else
-		lua_pushliteral(L, "open");
-	return 1;
-}
-
-/* }}} */
-
 void
 box_lua_misc_init(struct lua_State *L)
 {
@@ -590,8 +527,6 @@ box_lua_misc_init(struct lua_State *L)
 		{"prepare_auth", lbox_prepare_auth},
 		{"select", lbox_select},
 		{"txn_set_isolation", lbox_txn_set_isolation},
-		{"read_view_list", lbox_read_view_list},
-		{"read_view_status", lbox_read_view_status},
 		{"generate_space_id", lbox_generate_space_id},
 		{"generate_func_id", lbox_generate_func_id},
 		{"memtx_tx_gc", lbox_memtx_tx_gc},
